requirements.txt
streamlit
openai
azure-identity
PyPDF2
BioPython
# Web Search & Content
duckduckgo-search
requests
# PDF OCR & Processing (OCR REMOVED)
# pytesseract # REMOVED
# Pillow # REMOVED
# pdf2image # REMOVED
# Vector DB & Embeddings
langchain
langchain-community
sentence-transformers
faiss-cpu
# Streamlit Extras (Optional - for copy button)
streamlit-extras
# Word document export
python-docx

# Environment variable management
python-dotenv

# pdfplumber # Retaining as a comment from previous version, not actively used.

streamlit-option-menu
streamlit-lottie
streamlit_js_eval